Fix the registration of dynamic IDs.
This commit papers over an ugly problem: loop class first calculates
all the iterations, and only then presents returns the values as a
list. This is a problem if any of the internal parameters try to refer
to the loop object, e.g. to access other invocations. This becomes
a problem with dynamic IDs, because deferred calculation is impossible
if such references appear in an ID expression.

A better fix would be to make loop not a caching parameter and provide
access to loop iterations directly.