Prevent recursion into the same node during ASTTypeVisitor's work
A null (==unknown type) is places in the cache when recursing downwards from a node.
If any of the subexpression refer to this node, they will retrieve type==null from the cache.
The null replacement object in the cache gets its own static internal class, so that
it displays properly in debug messages (getSimpleName() of inline classes returns an
empty string, yuck).

A bug which would cause the null replacement is fixes in putCached().