JavaClassGen: change the names of cast/make methods to better reflect meaning
Previously, the whole set of methods to cast or not cast depending on expected
type was broken, because the expected type was always null, since type guessing
wasn't really functional. When now it is better, the generated code would not compile,
because the (expected) dynamic types were different then the (known) static types.

The whole subject of what type annotations mean should be revisitied.