Remove IndexClass and generate the whole class [generated code broken]
The efficiency of loop variable access is not so important (signalml loops
are likely to be small, over some header or something) and it should be easier
to reuse the normal Param<Type> machinery.

Creating type-specific code is only possible when IndexClass can be specialized,
so the stub implementation doesn't make much sense.