specs: define header fields directly as strings
We do not want to have to evalute the fields using an expression
parser. For example, a tool to pick codecs should not have to
implement them to extract basic information. Also, the fields
should be constant, and always of type string. So let's make
them constant string fields, specified without quoting.

Unused (but present in various specfiles) field <text_info> is
converted to header/format_id/description.